+++ lib/kitchen/directions/bake_first_elements.rb
@@ -4,6 +4,7 @@ module Kitchen
   module Directions
     module BakeFirstElements
       def self.v1(within:)
+        # add has-first-element class
         selectors = [
           'div.os-problem-container > div.os-table',
           'div.os-problem-container > span[data-type="media"]',
@@ -16,6 +17,13 @@ module Kitchen
             problem.parent.add_class('has-first-element')
           end
         end
+
+        # add first-inline-element class
+        inline_list = within&.first('div.os-solution-container > ol[type="1"]:first-child')
+        return unless inline_list
+
+        inline_list.add_class('first-inline-list-element')
+        inline_list.parent.add_class('has-first-inline-list-element')
       end
     end
   end